Update to handle listening for reactions on the LFG embed. Created LFGConstants for reusable constants.
